+++ tests/test_benchmarks/standard_benchmarks.py
@@ -1,0 +1,16 @@
+import nnbench
+
+
+@nnbench.benchmark(tags=("standard",))
+def double(x: int) -> int:
+    return x * 2
+
+
+@nnbench.benchmark(tags=("standard",))
+def triple(y: int) -> int:
+    return y * 3
+
+
+@nnbench.benchmark(tags=("standard",))
+def prod(x: int, y: int) -> int:
+    return x * y
